Move: SchoolButtonsコンポーネントをfeatures配下に移動させました。
