+++ app/javascript/components/features/SchoolButtons.js
@@ -1,5 +1,5 @@
 import React, { useState } from 'react';
-import Button from './Button';
+import Button from '../components/Button';
 
 const SchoolButtons = ({ identifier, selectedButton, onChange }) => {
